pyproject.toml
[project]
name = "splunk-mcp-template"
version = "0.1.0"
description = "Template MCP to run Splunk searches, analyze thresholds, and open Jira tickets."
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[project.scripts]
splunk-mcp = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

